Flawless linting
