+++ flathunter/web/__init__.py
@@ -1,4 +1,5 @@
 """Initialize website module"""
+# pylint: disable=wrong-import-position
 from flask import Flask
 from flask.logging import create_logger
