+++ setup.py
@@ -1,3 +1,4 @@
+"""Setup file for flathunters (used for compatibility with 'pip install -e .'"""
 from setuptools import setup, find_packages
 
 setup(name="flathunter", packages=find_packages())
